requirements.txt
langchain_community==0.3.24
langchain_openai==0.3.21
langchain_qdrant==0.2.0
langchain_text_splitters==0.3.8
openai==1.84.0
protobuf==6.31.1
python-dotenv==1.1.0
Requests==2.32.3
streamlit==1.45.1
tiktoken==0.9.0
